Fixed the issue in tablet mode, when the fullscreen media viewer was opening only on the left side (the chat view) and not the chatlist view
