+++ presentation/src/main/java/org/monogram/presentation/features/profile/ProfileContent.kt
@@ -355,163 +355,6 @@ fun ProfileContent(component: ProfileComponent) {
             }
         }
 
-        val notImplemented = stringResource(R.string.not_implemented)
-        AnimatedVisibility(
-            visible = state.fullScreenImages != null,
-            enter = fadeIn() + scaleIn(initialScale = 0.9f),
-            exit = fadeOut() + scaleOut(targetScale = 0.9f)
-        ) {
-            state.fullScreenImages?.let { images ->
-                Box(modifier = Modifier.fillMaxSize()) {
-                    ImageViewer(
-                        images = images,
-                        startIndex = state.fullScreenStartIndex,
-                        onDismiss = component::onDismissViewer,
-                        autoDownload = false,
-                        downloadUtils = component.downloadUtils,
-                        onPageChanged = { index ->
-
-                            if (!state.isViewingProfilePhotos && state.canLoadMoreMedia && !state.isLoadingMoreMedia &&
-                                index >= images.size - 5
-                            ) {
-                                component.onLoadMoreMedia()
-                            }
-
-                            if (!state.isViewingProfilePhotos) {
-                                val photoMessages = state.mediaMessages.filter { it.content is MessageContent.Photo }
-
-                                val message = photoMessages.getOrNull(index)
-                                if (message != null) {
-                                    component.onDownloadMedia(message)
-
-                                    val nextMessage = photoMessages.getOrNull(index + 1)
-                                    if (nextMessage != null) {
-                                        component.onDownloadMedia(nextMessage)
-                                    }
-                                }
-                            }
-                        },
-                        onForward = { Toast.makeText(context, notImplemented, Toast.LENGTH_SHORT).show() },
-                        onDelete = { Toast.makeText(context, notImplemented, Toast.LENGTH_SHORT).show() },
-                        onCopyLink = { Toast.makeText(context, notImplemented, Toast.LENGTH_SHORT).show() },
-                        onCopyText = { Toast.makeText(context, notImplemented, Toast.LENGTH_SHORT).show() },
-                        captions = state.fullScreenCaptions.filterNotNull(),
-                        showImageNumber = false
-                    )
-
-                    if (state.isViewingProfilePhotos && state.isProfilePhotoHdLoading) {
-                        Surface(
-                            modifier = Modifier
-                                .align(Alignment.TopCenter)
-                                .padding(top = 56.dp),
-                            shape = RoundedCornerShape(14.dp),
-                            color = MaterialTheme.colorScheme.surfaceContainerHigh.copy(alpha = 0.92f)
-                        ) {
-                            Row(
-                                modifier = Modifier.padding(horizontal = 14.dp, vertical = 8.dp),
-                                verticalAlignment = Alignment.CenterVertically,
-                                horizontalArrangement = Arrangement.spacedBy(10.dp)
-                            ) {
-                                LoadingIndicator(
-                                    modifier = Modifier.size(16.dp)
-                                )
-                                Text(
-                                    text = "Loading HD",
-                                    style = MaterialTheme.typography.labelLarge,
-                                    color = MaterialTheme.colorScheme.onSurface
-                                )
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        AnimatedVisibility(
-            visible = state.fullScreenVideoPath != null,
-            enter = fadeIn() + scaleIn(initialScale = 0.9f),
-            exit = fadeOut() + scaleOut(targetScale = 0.9f)
-        ) {
-            state.fullScreenVideoPath?.let { path ->
-                val msg = state.mediaMessages.find {
-                    when (val content = it.content) {
-                        is MessageContent.Video -> content.path == path
-                        is MessageContent.Gif -> content.path == path
-                        is MessageContent.VideoNote -> content.path == path
-                        else -> false
-                    }
-                }
-                val videoContent = msg?.content as? MessageContent.Video
-                val fileId = videoContent?.fileId ?: (msg?.content as? MessageContent.Gif)?.fileId
-                ?: (msg?.content as? MessageContent.VideoNote)?.fileId ?: 0
-                val supportsStreaming = videoContent?.supportsStreaming ?: false
-
-                VideoViewer(
-                    path = path,
-                    onDismiss = component::onDismissViewer,
-                    isGesturesEnabled = true,
-                    isDoubleTapSeekEnabled = true,
-                    seekDuration = 10,
-                    isZoomEnabled = true,
-                    downloadUtils = component.downloadUtils,
-                    onForward = { Toast.makeText(context, notImplemented, Toast.LENGTH_SHORT).show() },
-                    onDelete = { Toast.makeText(context, notImplemented, Toast.LENGTH_SHORT).show() },
-                    onCopyLink = { Toast.makeText(context, notImplemented, Toast.LENGTH_SHORT).show() },
-                    onCopyText = { Toast.makeText(context, notImplemented, Toast.LENGTH_SHORT).show() },
-                    caption = state.fullScreenVideoCaption,
-                    fileId = fileId,
-                    supportsStreaming = supportsStreaming
-                )
-            }
-        }
-
-        AnimatedVisibility(
-            visible = state.miniAppUrl != null,
-            enter = slideInVertically(initialOffsetY = { it }) + fadeIn(),
-            exit = slideOutVertically(targetOffsetY = { it }) + fadeOut()
-        ) {
-            if (state.miniAppUrl != null && state.miniAppName != null) {
-                MiniAppViewer(
-                    baseUrl = state.miniAppUrl.toString(),
-                    botName = title,
-                    onDismiss = { component.onDismissMiniApp() },
-                    chatId = state.chatId,
-                    botUserId = state.user!!.id,
-                    webAppRepository = component.messageRepository,
-                )
-            }
-        }
-
-        AnimatedVisibility(
-            visible = state.isStatisticsVisible,
-            enter = slideInVertically(initialOffsetY = { it }) + fadeIn(),
-            exit = slideOutVertically(targetOffsetY = { it }) + fadeOut()
-        ) {
-            if (state.statistics != null) {
-                StatisticsViewer(
-                    title = stringResource(R.string.statistics_title),
-                    data = state.statistics!!,
-                    onDismiss = component::onDismissStatistics,
-                    onLoadGraph = component::onLoadStatisticsGraph
-                )
-            }
-        }
-
-        AnimatedVisibility(
-            visible = state.isRevenueStatisticsVisible,
-            enter = slideInVertically(initialOffsetY = { it }) + fadeIn(),
-            exit = slideOutVertically(targetOffsetY = { it }) + fadeOut()
-        ) {
-            if (state.revenueStatistics != null) {
-                StatisticsViewer(
-                    title = stringResource(R.string.revenue_title),
-                    data = state.revenueStatistics!!,
-                    onDismiss = component::onDismissStatistics,
-                    onLoadGraph = component::onLoadStatisticsGraph
-                )
-            }
-        }
-
         ProfileQRDialog(
             state = state,
             onDismiss = component::onDismissQRCode
@@ -664,13 +507,6 @@ fun ProfileContent(component: ProfileComponent) {
                 state = state,
                 onDismiss = component::onDismissTOS,
                 onAccept = component::onAcceptTOS
-            )
-        }
-
-        state.selectedLocation?.let { location ->
-            LocationViewer(
-                location = location,
-                onDismiss = component::onDismissLocation
             )
         }
     }
